requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
django-countries==7.6.1
django-country==0.1.0
django-crispy-forms==2.1
djangorestframework==3.15.1
idna==3.7
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.5.0
stripe==9.3.0
typing_extensions==4.11.0
urllib3==2.2.1
